Migrate remainder of AssetGroup tests
[INTERNAL_BRANCH=sean/refactor-pipeline-symbol-renames]
[DAGSTER_CHECKOUT_DEPTH=1000]
